move AbstractEnumTypeAdapter<T> to cloud.grabsky.configuration.adapter package
